diagnose: split .svelte.js dev cases across owning specs
Replace combined module_compilation_dev case with three focused cases:
- module_dev_state_tag → state-rune.md
- module_dev_derived_tag → derived-state.md
- module_dev_console_log_wrap → unknown.md (suggested dev-console-instrumentation.md)

unknown.md additionally tracks the shared cross-cutting root cause: compile_module
does not thread dev into the codegen-side transform pipeline.
