+++ tasks/compiler_tests/cases2/module_dev_console_log_wrap/case-rust.js
@@ -1,0 +1,8 @@
+import * as $ from "svelte/internal/client";
+let count = $.state(0);
+export function increment() {
+	$.update(count);
+}
+$.user_effect(() => {
+	console.log("count changed:", $.get(count));
+});
