+++ tasks/compiler_tests/cases2/module_dev_console_log_wrap/case.svelte.js
@@ -1,0 +1,9 @@
+let count = $state(0);
+
+export function increment() {
+	count++;
+}
+
+$effect(() => {
+	console.log("count changed:", count);
+});
