+++ tasks/compiler_tests/cases2/module_dev_console_log_wrap/config.json
