+++ tasks/compiler_tests/cases2/module_dev_derived_tag/case-rust.js
@@ -1,15 +1,9 @@
 import * as $ from "svelte/internal/client";
 let count = $.state(0);
 const doubled = $.derived(() => $.get(count) * 2);
-export function increment() {
-	$.update(count);
-}
-export function getCount() {
-	return $.get(count);
-}
 export function getDoubled() {
 	return $.get(doubled);
 }
-$.user_effect(() => {
-	console.log("count changed:", $.get(count));
-});
+export function increment() {
+	$.update(count);
+}
