+++ tasks/compiler_tests/cases2/module_dev_derived_tag/case-svelte.js
@@ -1,0 +1,9 @@
+import * as $ from "svelte/internal/client";
+let count = $.tag($.state(0), "count");
+const doubled = $.tag($.derived(() => $.get(count) * 2), "doubled");
+export function getDoubled() {
+	return $.get(doubled);
+}
+export function increment() {
+	$.update(count);
+}
