+++ tasks/compiler_tests/cases2/module_dev_derived_tag/case.svelte.js
@@ -1,0 +1,10 @@
+let count = $state(0);
+const doubled = $derived(count * 2);
+
+export function getDoubled() {
+	return doubled;
+}
+
+export function increment() {
+	count++;
+}
