+++ tasks/compiler_tests/cases2/module_dev_state_tag/case-svelte.js
@@ -1,0 +1,8 @@
+import * as $ from "svelte/internal/client";
+let count = $.tag($.state(0), "count");
+export function increment() {
+	$.update(count);
+}
+export function getCount() {
+	return $.get(count);
+}
