+++ tasks/compiler_tests/cases2/module_dev_state_tag/case.svelte.js
@@ -1,0 +1,9 @@
+let count = $state(0);
+
+export function increment() {
+	count++;
+}
+
+export function getCount() {
+	return count;
+}
