+++ tasks/compiler_tests/cases2/module_dev_state_tag/config.json
@@ -1,0 +1,4 @@
+{
+  "dev": true,
+  "filename": "input.svelte.js"
+}
